requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.0
h11==0.12.0
pydantic==1.8.2
PyJWT==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
